Factor the query queue to a separate TU
Also rename from `pending_queue` to `query_queue` and document the
exposed functions.
